+++ integrador/src/main/java/com/backend/digitalhouse/integrador/dto/entrada/modificacion/PacienteModificacionEntradaDto.java
@@ -9,7 +9,7 @@ import java.time.LocalDate;
 public class PacienteModificacionEntradaDto {
 
     @NotNull
-    private int id;
+    private Long id;
 
     @NotNull
     private String nombre;
@@ -29,7 +29,7 @@ public class PacienteModificacionEntradaDto {
     public PacienteModificacionEntradaDto() {
     }
 
-    public PacienteModificacionEntradaDto(int id, String nombre, String apellido, String dni, LocalDate fechaIngreso, DomicilioModificacionEntradaDto domicilio) {
+    public PacienteModificacionEntradaDto(Long id, String nombre, String apellido, String dni, LocalDate fechaIngreso, DomicilioModificacionEntradaDto domicilio) {
         this.id = id;
         this.nombre = nombre;
         this.apellido = apellido;
@@ -38,11 +38,11 @@ public class PacienteModificacionEntradaDto {
         this.domicilio = domicilio;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
