+++ integrador/src/main/java/com/backend/digitalhouse/integrador/dto/salida/turno/OdontologoTurnoSalidaDto.java
@@ -3,24 +3,24 @@ package com.backend.digitalhouse.integrador.dto.salida.turno;
 public class OdontologoTurnoSalidaDto {
 
 
-    private int id;
+    private Long id;
     private String nombre;
     private String apellido;
 
     public OdontologoTurnoSalidaDto() {
     }
 
-    public OdontologoTurnoSalidaDto(int id, String nombre, String apellido) {
+    public OdontologoTurnoSalidaDto(Long id, String nombre, String apellido) {
         this.id = id;
         this.nombre = nombre;
         this.apellido = apellido;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
